+++ GUI/legacy/src/main/java/com/byrdofafeather/Predictors.java
