+++ requirements.txt
@@ -1,7 +1,6 @@
 requests==2.18.4
-nltk==3.2.5
 numpy==1.13.3
-scikit-learn==0.19.1
+sklearn==0.19.1
 tensorflow==1.3.0
 pandas==0.21.0
 matplotlib==2.1.1
